fix(grcp-geo-types): add support for use with utoipa
