+++ client-grpc/tests/resources/vehicle.rs
@@ -26,9 +26,7 @@ pub async fn scenario(
     // Insert vehicles for each mock object
     for vehicle_data in data {
         println!("Starting insert vehicle");
-        let result = client
-            .insert(tonic::Request::new(vehicle_data.clone()))
-            .await;
+        let result = client.insert(vehicle_data.clone()).await;
 
         let expected = get_log_string("insert", name);
         println!("expected message: {}", expected);
@@ -58,7 +56,7 @@ pub async fn scenario(
     };
 
     // Check if all vehicles can be retrieved from the backend
-    let result = client.search(tonic::Request::new(not_deleted_filter)).await;
+    let result = client.search(not_deleted_filter).await;
     let expected = get_log_string("search", name);
     println!("expected message: {}", expected);
     assert!(logger.any(|log| check_log_string_matches(log, &expected)));
@@ -72,9 +70,9 @@ pub async fn scenario(
 
     // Check if we can get a single vehicle based on their id
     let result = client
-        .get_by_id(tonic::Request::new(Id {
+        .get_by_id(Id {
             id: vehicle_id.clone(),
-        }))
+        })
         .await;
 
     let expected = get_log_string("get_by_id", name);
@@ -88,9 +86,9 @@ pub async fn scenario(
 
     // Check if we can delete the vehicle
     let result = client
-        .delete(tonic::Request::new(Id {
+        .delete(Id {
             id: vehicle_id.clone(),
-        }))
+        })
         .await;
 
     let expected = get_log_string("delete", name);
