+++ client-grpc/tests/resources/vertipad.rs
@@ -26,9 +26,7 @@ pub async fn scenario(
     // Insert vertipads for each mock object
     for vertipad_data in data {
         println!("Starting insert vertipad");
-        let result = client
-            .insert(tonic::Request::new(vertipad_data.clone()))
-            .await;
+        let result = client.insert(vertipad_data.clone()).await;
 
         let expected = get_log_string("insert", name);
         println!("expected message: {}", expected);
@@ -55,7 +53,7 @@ pub async fn scenario(
     };
 
     // Check if all vertipads can be retrieved from the backend
-    let result = client.search(tonic::Request::new(not_deleted_filter)).await;
+    let result = client.search(not_deleted_filter).await;
     let expected = get_log_string("search", name);
     println!("expected message: {}", expected);
     assert!(logger.any(|log| check_log_string_matches(log, &expected)));
@@ -69,9 +67,9 @@ pub async fn scenario(
 
     // Check if we can get a single vertipad based on their id
     let result = client
-        .get_by_id(tonic::Request::new(Id {
+        .get_by_id(Id {
             id: vertipad_id.clone(),
-        }))
+        })
         .await;
 
     let expected = get_log_string("get_by_id", name);
@@ -85,9 +83,9 @@ pub async fn scenario(
 
     // Check if we can delete the vertipad
     let result = client
-        .delete(tonic::Request::new(Id {
+        .delete(Id {
             id: vertipad_id.clone(),
-        }))
+        })
         .await;
 
     let expected = get_log_string("delete", name);
