+++ server/src/grpc/macros.rs
@@ -53,7 +53,7 @@ macro_rules! build_grpc_server_link_service_impl {
                 &self,
                 request: Request<$link_other_resource>,
             ) -> Result<tonic::Response<()>, Status> {
-                grpc_warn!("(link) {} server.", self.get_name());
+                grpc_info!("(link) {} server.", self.get_name());
                 grpc_debug!("(link) request: {:?}", request);
                 let data: $link_other_resource = request.into_inner();
                 self.generic_link::<ResourceObject<$other_resource::Data>>(data.id.clone(), data.get_other_ids().try_into()?, false)
@@ -120,7 +120,7 @@ macro_rules! build_grpc_server_link_service_impl {
                 &self,
                 request: Request<$link_other_resource>,
             ) -> Result<tonic::Response<()>, Status> {
-                grpc_warn!("(replace_linked) {} server.", self.get_name());
+                grpc_info!("(replace_linked) {} server.", self.get_name());
                 grpc_debug!("(replace_linked) request: {:?}", request);
                 let data: $link_other_resource = request.into_inner();
                 self.generic_link::<ResourceObject<$other_resource::Data>>(data.id.clone(), data.get_other_ids().try_into()?, true)
@@ -178,7 +178,7 @@ macro_rules! build_grpc_server_link_service_impl {
             /// Returns [`tonic::Status`] with [`tonic::Code::NotFound`] if the provided `id` is not found in the database.
             #[cfg(not(feature = "stub_server"))]
             async fn unlink(&self, request: Request<Id>) -> Result<tonic::Response<()>, Status> {
-                grpc_warn!("(unlink) {} server.", self.get_name());
+                grpc_info!("(unlink) {} server.", self.get_name());
                 grpc_debug!("(unlink) request: {:?}", request);
                 self.generic_unlink(request).await
             }
@@ -228,7 +228,7 @@ macro_rules! build_grpc_server_link_service_impl {
                 &self,
                 request: Request<Id>,
             ) -> Result<tonic::Response<IdList>, Status> {
-                grpc_warn!("(get_linked_ids) {} server.", self.get_name());
+                grpc_info!("(get_linked_ids) {} server.", self.get_name());
                 grpc_debug!("(get_linked_ids) request: {:?}", request);
                 self.generic_get_linked_ids::<ResourceObject<$other_resource::Data>, $other_resource::Data>(request)
                     .await
@@ -258,7 +258,7 @@ macro_rules! build_grpc_server_link_service_impl {
                 &self,
                 request: Request<Id>,
             ) -> Result<tonic::Response<$other_resource::List>, Status> {
-                grpc_warn!("(get_linked) {} server.", self.get_name());
+                grpc_info!("(get_linked) {} server.", self.get_name());
                 grpc_debug!("(get_linked) request: {:?}", request);
                 self.generic_get_linked::<ResourceObject<$other_resource::Data>, $other_resource::Data, $other_resource::List>(
                     request,
@@ -343,6 +343,7 @@ macro_rules! grpc_server {
                 ".service.rs"
             ));
             pub use rpc_service_server::*;
+            pub use $crate::grpc::server::grpc_geo_types::*;
 
             #[doc = concat!(stringify!($resource), "module including mock file")]
 
@@ -473,7 +474,7 @@ macro_rules! grpc_server {
                 ///
                 /// A new [`Uuid`](uuid::Uuid) will be generated by the database and returned as `id` as part of the returned [`Object`].
                 ///
-                /// # Example
+                /// # Examples
                 /// ```
                 /// use svc_storage::resources::Id;
                 #[doc = concat!("use svc_storage::resources::", $rpc_string, "::{Data, Response, GrpcServer, RpcService};")]
@@ -539,7 +540,7 @@ macro_rules! grpc_server {
                 /// A field mask can be provided to restrict updates to specific fields.
                 /// Returns the updated [`Response`] on success.
                 ///
-                /// # Example
+                /// # Examples
                 /// ```
                 /// use svc_storage::resources::Id;
                 #[doc = concat!("use svc_storage::resources::", $rpc_string, "::{UpdateObject, Response, GrpcServer, RpcService};")]
@@ -607,7 +608,7 @@ macro_rules! grpc_server {
 
                 #[doc = concat!("Takes an [`Id`] to set the matching ", $rpc_string, " record as deleted in the database.")]
                 ///
-                /// # Example
+                /// # Examples
                 /// ```
                 /// use svc_storage::resources::Id;
                 #[doc = concat!("use svc_storage::resources::", $rpc_string, "::{GrpcServer, RpcService};")]
