+++ server/src/grpc/mod.rs
@@ -9,8 +9,6 @@ mod link_service;
 mod simple_service;
 
 pub use crate::common::ArrErr;
-use crate::resources::{GeoLineString, GeoPoint, GeoPolygon};
-use geo_types::{LineString, Point, Polygon};
 pub use link_service::GrpcLinkService;
 pub use simple_service::GrpcSimpleService;
 
@@ -20,6 +18,8 @@ use std::time::SystemTime;
 use std::{fmt::Debug, vec};
 use tokio::runtime::{Handle, Runtime};
 use tonic::Status;
+
+use server::grpc_geo_types::*;
 
 /// gRPC field types
 #[derive(Debug, Clone, PartialEq)]
@@ -318,6 +318,7 @@ impl From<GrpcFieldOption> for Option<GrpcField> {
 /// The handle can be used to spawn a separate task, or run an async function
 /// from a non-async function.
 ///
+/// # Examples
 /// ```
 /// use svc_storage::grpc::get_runtime_handle;
 /// use svc_storage::postgres::simple_resource::PsqlType;
