+++ server/src/resources/base/test_util.rs
@@ -31,11 +31,11 @@ pub struct TestData {
     pub u32_vec: ::prost::alloc::vec::Vec<u32>,
 
     #[prost(message, optional, tag = "10")]
-    pub geo_point: ::core::option::Option<crate::resources::GeoPoint>, // Always passed as an option, but will check for mandatory state
+    pub geo_point: ::core::option::Option<crate::resources::grpc_geo_types::GeoPoint>, // Always passed as an option, but will check for mandatory state
     #[prost(message, optional, tag = "11")]
-    pub geo_polygon: ::core::option::Option<crate::resources::GeoPolygon>, // Always passed as an option, but will check for mandatory state
+    pub geo_polygon: ::core::option::Option<crate::resources::grpc_geo_types::GeoPolygon>, // Always passed as an option, but will check for mandatory state
     #[prost(message, optional, tag = "12")]
-    pub geo_line_string: ::core::option::Option<crate::resources::GeoLineString>, // Always passed as an option, but will check for mandatory state
+    pub geo_line_string: ::core::option::Option<crate::resources::grpc_geo_types::GeoLineString>, // Always passed as an option, but will check for mandatory state
 
     #[prost(string, optional, tag = "21")]
     pub optional_string: ::core::option::Option<::prost::alloc::string::String>,
@@ -51,11 +51,12 @@ pub struct TestData {
     pub optional_uuid: ::core::option::Option<::prost::alloc::string::String>,
 
     #[prost(message, optional, tag = "30")]
-    pub optional_geo_point: ::core::option::Option<crate::resources::GeoPoint>,
+    pub optional_geo_point: ::core::option::Option<crate::resources::grpc_geo_types::GeoPoint>,
     #[prost(message, optional, tag = "31")]
-    pub optional_geo_polygon: ::core::option::Option<crate::resources::GeoPolygon>,
+    pub optional_geo_polygon: ::core::option::Option<crate::resources::grpc_geo_types::GeoPolygon>,
     #[prost(message, optional, tag = "32")]
-    pub optional_geo_line_string: ::core::option::Option<crate::resources::GeoLineString>,
+    pub optional_geo_line_string:
+        ::core::option::Option<crate::resources::grpc_geo_types::GeoLineString>,
 
     #[prost(string, optional, tag = "40")]
     pub read_only: ::core::option::Option<::prost::alloc::string::String>,
