+++ server/src/resources/vertipad/mod.rs
@@ -13,7 +13,7 @@ use super::base::simple_resource::*;
 use super::base::{FieldDefinition, ResourceDefinition};
 use crate::common::ArrErr;
 use crate::grpc::{GrpcDataObjectType, GrpcField, GrpcFieldOption};
-use crate::resources::GeoPoint;
+use crate::resources::grpc_geo_types::GeoPoint;
 
 // Generate `From` trait implementations for GenericResource into and from Grpc defined Resource
 crate::build_generic_resource_impl_from!();
